+++ src/controllers/feedback.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import Feedback from "../models/feedback.model";
 import logger from "../libs/logger";
 import { ExtendedRequest } from "../interfaces/express";
+import RedisService from "../services/redis.service";
 
 const addFeedback = async (req: ExtendedRequest, res: Response) => {
   const { from_user, to_user, content, company_name } = req.body;
@@ -12,6 +13,7 @@ const addFeedback = async (req: ExtendedRequest, res: Response) => {
       content,
       company_name,
     });
+    await RedisService.clearCache(`cv${to_user}`);
     return res.send(feedback);
   } catch (err) {
     logger.error(err);
@@ -41,12 +43,14 @@ const updateFeedback = async (req: ExtendedRequest, res: Response) => {
   const { id } = req.params;
   const update = req.body;
   const feedback = await Feedback.findOne({ where: { id } });
+  const { to_user } = feedback;
   if (!feedback) {
     logger.error("No Feedback found");
     res.status(404).send("An Error Occured");
   }
   try {
     const newFeedback = await Feedback.update(update, { where: { id } });
+    await RedisService.clearCache(`cv${to_user}`);
     res.send(newFeedback);
   } catch (err) {
     logger.error(err.message);
@@ -57,7 +61,12 @@ const updateFeedback = async (req: ExtendedRequest, res: Response) => {
 const deleteFeedback = async (req: ExtendedRequest, res: Response) => {
   const { id } = req.params;
   try {
-    await Feedback.destroy({ where: { id } });
+    const result = await Feedback.findOne({ where: { id } });
+    const { to_user } = result;
+    if (result) {
+      await Feedback.destroy({ where: { id } });
+      await RedisService.clearCache(`cv${to_user}`);
+    }
     return res.send("Successfully Exterminated");
   } catch (err) {
     logger.error(err.message);
