+++ src/controllers/projects.ts
@@ -3,6 +3,7 @@ import { Response } from "express";
 import Project from "../models/project.model";
 import logger from "../libs/logger";
 import { ExtendedRequest } from "../interfaces/express";
+import RedisService from "../services/redis.service";
 
 const createProject = async (req: ExtendedRequest, res: Response) => {
   const { user_id, description } = req.body;
@@ -13,7 +14,7 @@ const createProject = async (req: ExtendedRequest, res: Response) => {
       description,
       image,
     });
-
+    await RedisService.clearCache(`cv${user_id}`);
     res.json({
       project,
     });
@@ -50,6 +51,8 @@ const updateOneProject = async (req: ExtendedRequest, res: Response) => {
   }
   try {
     const newProject = await Project.update(update, { where: { id } });
+    const { user_id } = project;
+    await RedisService.clearCache(`cv${user_id}`);
     res.send(newProject);
   } catch (err) {
     logger.error(err.message);
@@ -60,8 +63,13 @@ const updateOneProject = async (req: ExtendedRequest, res: Response) => {
 const deleteProject = async (req: ExtendedRequest, res: Response) => {
   const { id } = req.params;
   try {
-    await Project.destroy({ where: { id } });
-    return res.send("Successfully Exterminated");
+    const result = await Project.findOne({ where: { id } });
+    const { user_id } = result;
+    if (result) {
+      await Project.destroy({ where: { id } });
+      await RedisService.clearCache(`cv${user_id}`);
+      return res.send("Successfully Exterminated");
+    }
   } catch (err) {
     logger.error(err.message);
     return err;
